Refactor donation helpers into shared ES module
