+++ donate.html
@@ -47,7 +47,7 @@
   </main>
   <div id="footer"></div>
   <script src="js/config.js"></script>
-  <script src="js/donate.js"></script>
+  <script type="module" src="js/donate.js"></script>
   <script src="js/header.js"></script>
   <script src="js/footer.js"></script>
 </body>
